[Refactor] 마이페이지 코디네이터 통합
