+++ MateRunner/MateRunner/Application/AppDelegate.swift
@@ -130,8 +130,9 @@ extension AppDelegate : UNUserNotificationCenterDelegate {
               let tabBarCoordinator = appCoordinator.findCoordinator(type: .tab) as? TabBarCoordinator,
               let myPageCoordinator = appCoordinator.findCoordinator(type: .mypage) as? MyPageCoordinator else { return }
         tabBarCoordinator.selectPage(.mypage)
-        // TODO: push NotificationViewController after integrates mypage subcoordinators
-        myPageCoordinator.showNotificationFlow()
+        guard (myPageCoordinator.navigationController.viewControllers.last
+               is NotificationViewController == false) else { return }
+        myPageCoordinator.pushNotificationViewController()
     }
     
     private func configureInvitation(with userInfo: [AnyHashable: Any]) {
