--- MateRunner/MateRunner/Presentation/MyPageScene/Coordinator/Protocol/LicenseCoordinator.swift
@@ -1,12 +1,0 @@
-//
-//  LicenseCoordinator.swift
-//  MateRunner
-//
-//  Created by 김민지 on 2021/11/24.
-//
-
-import Foundation
-
-protocol LicenseCoordinator: Coordinator {
-    func pushLicenseViewController()
-}
