--- MateRunner/MateRunner/Presentation/MyPageScene/Coordinator/Protocol/NotificationCoordinator.swift
@@ -1,12 +1,0 @@
-//
-//  NotificationCoordinator.swift
-//  MateRunner
-//
-//  Created by 김민지 on 2021/11/24.
-//
-
-import Foundation
-
-protocol NotificationCoordinator: Coordinator {
-    func pushNotificationViewController()
-}
