--- MateRunner/MateRunner/Presentation/MyPageScene/Coordinator/Protocol/ProfileEditCoordinator.swift
@@ -1,12 +1,0 @@
-//
-//  ProfileEditCoordinator.swift
-//  MateRunner
-//
-//  Created by 김민지 on 2021/11/24.
-//
-
-import Foundation
-
-protocol ProfileEditCoordinator: Coordinator {
-    func pushProfileEditViewController(with nickname: String)
-}
